Move more code out of search effect
